+++ internal/cli/integration/conflict_test.go
@@ -1,0 +1,65 @@
+package integration
+
+import (
+	"testing"
+)
+
+// =============================================================================
+// Conflict Resolution Integration Tests
+//
+// These tests cover scenarios where rebasing causes conflicts and the user
+// must resolve them using `stackit continue`.
+// =============================================================================
+
+func TestConflictResolution(t *testing.T) {
+	t.Parallel()
+	binaryPath := getStackitBinary(t)
+
+	t.Run("continue through cascading conflicts in stack", func(t *testing.T) {
+		// TODO: Implement multi-level conflict resolution test
+		//
+		// Scenario:
+		// 1. Build stack: main → branch-a → branch-b → branch-c
+		//    where each branch modifies the SAME file (to create conflicts)
+		// 2. Add a new commit to main that modifies the same file
+		// 3. Run `stackit restack` from branch-a
+		// 4. Verify: restack stops at branch-a with conflict
+		// 5. Resolve conflict, run `stackit continue`
+		// 6. Verify: restack continues, stops at branch-b with conflict
+		// 7. Resolve conflict, run `stackit continue`
+		// 8. Verify: restack continues, stops at branch-c with conflict
+		// 9. Resolve conflict, run `stackit continue`
+		// 10. Verify: all branches are now successfully restacked
+		//
+		// This tests:
+		// - Continuation state persistence across multiple conflicts
+		// - Proper resumption of restack after conflict resolution
+		// - Cascading conflicts through a deep stack
+
+		t.Skip("TODO: Implement cascading conflict resolution test")
+
+		_ = binaryPath // Will be used when implemented
+	})
+
+	t.Run("continue preserves stack structure after mid-stack conflict", func(t *testing.T) {
+		// TODO: Implement mid-stack conflict preservation test
+		//
+		// Scenario:
+		// 1. Build stack: main → a → b → c → d
+		// 2. Amend branch-b with conflicting changes
+		// 3. Run `stackit restack --upstack` from branch-b
+		// 4. Verify: conflict occurs at branch-c
+		// 5. Resolve conflict, run continue
+		// 6. Verify: branch-c and branch-d are properly restacked
+		// 7. Verify: parent relationships are preserved correctly
+		//
+		// This tests:
+		// - Restack from mid-stack position
+		// - Conflict in middle of upstack chain
+		// - Preservation of deep stack structure
+
+		t.Skip("TODO: Implement mid-stack conflict preservation test")
+
+		_ = binaryPath // Will be used when implemented
+	})
+}
