+++ internal/cli/integration/split_test.go
@@ -1,0 +1,81 @@
+package integration
+
+import (
+	"testing"
+)
+
+// =============================================================================
+// Split Workflow Integration Tests
+//
+// These tests cover the split command which extracts files from a branch
+// into a new parent branch, then restacks all affected branches.
+// =============================================================================
+
+func TestSplitWorkflow(t *testing.T) {
+	t.Parallel()
+	binaryPath := getStackitBinary(t)
+
+	t.Run("split mid-stack branch with multiple children restacks all descendants", func(t *testing.T) {
+		// TODO: Implement split with multiple children test
+		//
+		// Scenario - Diamond structure with split:
+		//
+		// Before split:
+		//           main
+		//             |
+		//         feature-a (has files: config.go, api.go, utils.go)
+		//          /     \
+		//      child-1  child-2
+		//                  |
+		//              grandchild
+		//
+		// After split --by-file config.go,api.go:
+		//
+		//           main
+		//             |
+		//        feature-a_split (has: config.go, api.go)
+		//             |
+		//         feature-a (has: utils.go only)
+		//          /     \
+		//      child-1  child-2
+		//                  |
+		//              grandchild
+		//
+		// Verify:
+		// - New parent branch created with extracted files
+		// - Original branch only has remaining files
+		// - All 3 descendants (child-1, child-2, grandchild) are restacked
+		// - Parent relationships are updated correctly
+		//
+		// This tests:
+		// - Split creates correct parent branch
+		// - Multiple children are all restacked
+		// - Deep descendants (grandchild) are handled
+		// - Diamond structure is preserved
+
+		t.Skip("TODO: Implement split mid-stack with multiple children test")
+
+		_ = binaryPath // Will be used when implemented
+	})
+
+	t.Run("split at stack bottom updates all upstack branches", func(t *testing.T) {
+		// TODO: Implement split at bottom test
+		//
+		// Scenario:
+		// 1. Build: main → feature-a (4 files) → feature-b → feature-c
+		// 2. Split feature-a to extract 2 files to new parent
+		// 3. Verify:
+		//    - New branch feature-a_split is between main and feature-a
+		//    - feature-a, feature-b, feature-c all properly restacked
+		//    - Commit history is clean
+		//
+		// This tests:
+		// - Split at the base of a deep stack
+		// - All upstack branches restack correctly
+		// - No orphaned commits
+
+		t.Skip("TODO: Implement split at stack bottom test")
+
+		_ = binaryPath // Will be used when implemented
+	})
+}
